+++ Data/Save/SaveFile01.json
@@ -1,1 +1,1 @@
-[[3,4294967295,0,1],[2,0,{"position":{"x":3.4300310611724854,"y":0.0,"z":3.8600406646728516},"rotation":{"w":1.0,"x":0.0,"y":0.0,"z":0.0},"scale":{"x":1.0,"y":1.0,"z":1.0}},1,{"position":{"x":0.0,"y":0.0,"z":0.0},"rotation":{"w":1.0,"x":0.0,"y":0.0,"z":0.0},"scale":{"x":1.0,"y":1.0,"z":1.0}}],[2,0,{"GUID":"6bfcb608-1334-4e88-b045-edc6c04c778d","Name":"FakeCube.test"},1,{"GUID":"927a7ff3-9e75-4089-be76-223599d1a89b","Name":"FakeCube.test"}],[1,0],[0]]
+[[3,4294967295,0,1],[2,0,{"position":{"x":-2.370006799697876,"y":0.0,"z":0.5299995541572571},"rotation":{"w":1.0,"x":0.0,"y":0.0,"z":0.0},"scale":{"x":1.0,"y":1.0,"z":1.0}},1,{"position":{"x":0.0,"y":0.0,"z":0.0},"rotation":{"w":1.0,"x":0.0,"y":0.0,"z":0.0},"scale":{"x":1.0,"y":1.0,"z":1.0}}],[2,0,{"GUID":"e1409f0b-7eab-4d5f-ab42-6c19ac0a9f99","Name":"FakeCube.test"},1,{"GUID":"b375c2f9-de44-45f5-8e31-4a6e22b3e601","Name":"FakeCube.test"}],[1,0],[0]]
